Cabin.glyphs: Fixed all errors reported in Preflight font script.
We now have a clean repo, 1 master .glyphs file with all masters and instances, correct vertical metrics, Better OT features. Luckily the original font was to a very very high standard. Most fonts are not this easy to work on.

I will need to repeat all the steps we did on the fonts to the Italics and Condensed files which existed in the old repo. I will also need to generate some tests fonts and run them through font bakery. I leave the font bakery step till the designers have finished working on the repo